feat: implement download functionality for different OS in SidebarUserProfile and DownloadButton components, enhance user experience with localized download messages
