+++ surfsense_web/components/layout/ui/sidebar/SidebarUserProfile.tsx
@@ -3,6 +3,7 @@
 import {
 	Check,
 	ChevronUp,
+	Download,
 	ExternalLink,
 	Info,
 	Languages,
@@ -29,6 +30,8 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { Spinner } from "@/components/ui/spinner";
 import { useLocaleContext } from "@/contexts/LocaleContext";
+import { usePlatform } from "@/hooks/use-platform";
+import { GITHUB_RELEASES_URL, usePrimaryDownload } from "@/lib/desktop-download-utils";
 import { APP_VERSION } from "@/lib/env-config";
 import { cn } from "@/lib/utils";
 import type { User } from "../../types/layout.types";
@@ -149,10 +152,13 @@ export function SidebarUserProfile({
 }: SidebarUserProfileProps) {
 	const t = useTranslations("sidebar");
 	const { locale, setLocale } = useLocaleContext();
+	const { isDesktop } = usePlatform();
+	const { os, primary } = usePrimaryDownload();
 	const [isLoggingOut, setIsLoggingOut] = useState(false);
 	const bgColor = stringToColor(user.email);
 	const initials = getInitials(user.email);
 	const displayName = user.name || user.email.split("@")[0];
+	const downloadUrl = primary?.url ?? GITHUB_RELEASES_URL;
 
 	const handleLanguageChange = (newLocale: "en" | "es" | "pt" | "hi" | "zh") => {
 		setLocale(newLocale);
@@ -293,6 +299,15 @@ export function SidebarUserProfile({
 								</DropdownMenuSubContent>
 							</DropdownMenuPortal>
 						</DropdownMenuSub>
+
+						{!isDesktop && (
+							<DropdownMenuItem asChild className="font-medium">
+								<a href={downloadUrl} target="_blank" rel="noopener noreferrer">
+									<Download className="h-4 w-4" strokeWidth={2.5} />
+									{t("download_for_os", { os })}
+								</a>
+							</DropdownMenuItem>
+						)}
 
 						<DropdownMenuSeparator className="dark:bg-neutral-700" />
 
@@ -438,6 +453,15 @@ export function SidebarUserProfile({
 							</DropdownMenuSubContent>
 						</DropdownMenuPortal>
 					</DropdownMenuSub>
+
+					{!isDesktop && (
+						<DropdownMenuItem asChild className="font-medium">
+							<a href={downloadUrl} target="_blank" rel="noopener noreferrer">
+								<Download className="h-4 w-4" strokeWidth={2.5} />
+								{t("download_for_os", { os })}
+							</a>
+						</DropdownMenuItem>
+					)}
 
 					<DropdownMenuSeparator className="dark:bg-neutral-700" />
