+++ surfsense_web/messages/hi.json
@@ -693,6 +693,7 @@
 		"learn_more": "और जानें",
 		"documentation": "दस्तावेज़ीकरण",
 		"github": "GitHub",
+		"download_for_os": "{os} के लिए डाउनलोड करें",
 		"inbox": "इनबॉक्स",
 		"search_inbox": "इनबॉक्स में खोजें",
 		"mark_all_read": "सभी पढ़ा हुआ चिह्नित करें",
