+++ surfsense_web/messages/pt.json
@@ -693,6 +693,7 @@
 		"learn_more": "Saiba mais",
 		"documentation": "Documentação",
 		"github": "GitHub",
+		"download_for_os": "Baixar para {os}",
 		"inbox": "Caixa de entrada",
 		"search_inbox": "Pesquisar caixa de entrada",
 		"mark_all_read": "Marcar tudo como lido",
